requirements.txt
runpod
torch
torchvision
transformers
diffusers
accelerate
trimesh
pygltflib
pillow
numpy
safetensors
huggingface_hub
